Move the TextureQuadData struct to the new Velaptor.OpenGL.GPUData namespace & folder
