+++ Testing/VelaptorTests/OpenGL/GPUData/TextureQuadDataTesting.cs
@@ -2,10 +2,9 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace VelaptorTests.OpenGL
+namespace VelaptorTests.OpenGL.GPUData
 {
     using System.Numerics;
-    using Velaptor.OpenGL;
     using Velaptor.OpenGL.GPUData;
     using Xunit;
