+++ Velaptor/InternalExtensionMethods.cs
@@ -17,7 +17,6 @@ namespace Velaptor
     using SixLabors.ImageSharp;
     using SixLabors.ImageSharp.PixelFormats;
     using Velaptor.Graphics;
-    using Velaptor.OpenGL;
     using Velaptor.OpenGL.GPUData;
     using NETColor = System.Drawing.Color;
     using NETRectF = System.Drawing.RectangleF;
